fix(CPhysical): maxhealth member type (#37)

+++ entities/CPhysical.hpp
@@ -21,7 +21,7 @@ namespace rage
         char gap198[232];
 	    float m_health; //0x0280
 	    char pad_0284[28];
-        uint32_t m_maxhealth; //0x02A0
+        float m_maxhealth; //0x02A0
         char gap2A4[4];
         uint64_t qword2A8;
         char gap2B0[72];
